requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
psycopg2-binary
passlib[bcrypt]
bcrypt>=4.0.0,<4.1.0
python-jose[cryptography]
PyPDF2
httpx
python-multipart
python-dotenv
diff-match-patch
pymed
